Dialer.RequireStartTLS to configure STARTTLS security
It is possible to use a man-in-the-middle attack to strip STARTTLS
security from an SMTP connection. Because all modern SMTP servers
support STARTTLS, we recommend that `Dialer.RequireStartTLS` be set to
`mail.MandatoryStartTLS`.

For backwards compatibility, this library defaults to
`mail.OpportunisticStartTLS`, which means that the client will
fallback to a cleartext connection if the SMTP server does not support
the STARTTLS extension.

This has no effect on SMTPS connections where TLS wraps the connection
itself.